requirements.txt
Django==3.1.5
djangorestframework==3.12.2
pytz==2020.5
requests==2.25.1
